XQuartz: Fix a bunch of compilation warnings about style (cherry picked from commit 54000bdcbca52a2de31f7c1a1147de6d8e9dbbb8)

+++ hw/xquartz/GL/visualConfigs.c
@@ -240,7 +240,7 @@ void setVisualConfigs(void) {
     }
 
     if (c - visualConfigs != numConfigs) {
-        FatalError("numConfigs calculation error in setVisualConfigs!  numConfigs is %d  i is %d\n", numConfigs, c - visualConfigs);
+        FatalError("numConfigs calculation error in setVisualConfigs!  numConfigs is %d  i is %d\n", numConfigs, (int)(c - visualConfigs));
     }
     
     freeGlCapabilities(&caps);
